+++ app/routes/videos/index.tsx
@@ -1,7 +1,7 @@
 import { useLoaderDataExtra } from "../../utils/loader-utils";
 import type { PageHandle } from "../../utils/page-handle";
+import { VideoListComponent } from "./_ui";
 import type { VideosLoaderData } from "./index.server";
-import { VideoListComponent } from "./index.utils";
 export { loader } from "./index.server";
 
 export const handle: PageHandle = {
